Refactor image handling in serializers and utilities; add tests for ChannelPostImage creation and removal

+++ backend/project/channel/utils.py
@@ -18,14 +18,9 @@ def remove_unused_images():
             continue
 
         # Delete file from storage
-        if img.image:
-            if CONFIG["is_local"]:
-                # Local storage
-                if os.path.isfile(img.image.path):
-                    os.remove(img.image.path)
-            else:
-                # S3 storage
-                img.image.delete(save=False)
+        os.remove(img.image.path) if CONFIG["is_local"] and os.path.isfile(
+            img.image.path
+        ) else img.image.delete(save=False)
 
         # Delete DB record
         img.delete()
